requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.6
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.6.0
